+++ src/ecosim/model/architecture/functions/ArchitectureFunction.java
@@ -2,6 +2,6 @@ package ecosim.model.architecture.functions;
 
 public interface ArchitectureFunction {
 
-	float evaluate(float parameter);
+	double evaluate(double parameter);
 	
 }
